Revert "Use full path for curl to prevent bugs or code"
This reverts commit 738e5d12dcff2a1be764c31530e7318d22096c60, pull request #57.

The $PATH environment variable exists specifically for the purpose of looking-
up binaries' locations in a system-independent manner. If $PATH is overwritten,
which the user has full discretion to do, then so be it. If the $PATH is
overwritten by compromise, then the system itself is already compromised and
this change prevents nothing.
Furthermore, many operating systems install packages and user-compiled software
into PREFIX=/usr/local, causing this commit to break this script.

Signed-off-by: Bryce Chidester <bryce@cobryce.com>