Update the login API JSON
This change is based on a JSON riot.im sends to the server to login into
it. The json riot.im sends it is different from the official JSON API
documentation. So I have found this doc where there is a proposal for a
change:

https://docs.google.com/document/d/1-6ZSSW5YvCGhVFDyD2QExAUAdpCWjccvJT5xiyTTG2Y/
edit#

There is another link where the docs are updated:
https://kamax.io/matrix/api/client_server/unstable.html#
post-matrix-client-r0-login

Fixes #113